fix(quiz): Fix security setup and incorrect reference to user id for quizzes [SCRUM-201]

+++ backend/src/main/java/stud/ntnu/backend/model/UserQuizAttempt.java
@@ -1,7 +1,13 @@
 package stud.ntnu.backend.model;
 
-import jakarta.persistence.*;
 import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -15,7 +21,7 @@ public class UserQuizAttempt {
     private Long id;
 
     @Column(name = "user_id", nullable = false)
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "quiz_id", nullable = false)
     private Long quizId;
